requirements.txt
transformers==4.46.3
tokenizers==0.20.3
einops
addict
easydict
fastapi
uvicorn[standard]
python-multipart
Pillow
pdf2image
